Create invoice endpoint
