+++ LBH.AdultSocialCare.Transactions.Api/V1/Controllers/InvoicesController.cs
@@ -1,14 +1,15 @@
 using LBH.AdultSocialCare.Transactions.Api.V1.Boundary.InvoiceBoundaries.Request;
 using LBH.AdultSocialCare.Transactions.Api.V1.Boundary.InvoiceBoundaries.Response;
+using LBH.AdultSocialCare.Transactions.Api.V1.Boundary.InvoiceBoundary.Response;
 using LBH.AdultSocialCare.Transactions.Api.V1.Exceptions.Models;
 using LBH.AdultSocialCare.Transactions.Api.V1.Factories;
 using LBH.AdultSocialCare.Transactions.Api.V1.UseCase.InvoiceUseCases.Interfaces;
+using LBH.AdultSocialCare.Transactions.Api.V1.UseCase.PayRunUseCases.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using LBH.AdultSocialCare.Transactions.Api.V1.UseCase.PayRunUseCases.Interfaces;
 
 namespace LBH.AdultSocialCare.Transactions.Api.V1.Controllers
 {
@@ -22,12 +23,15 @@ namespace LBH.AdultSocialCare.Transactions.Api.V1.Controllers
         private readonly IGetInvoiceItemPaymentStatusesUseCase _getInvoiceItemPaymentStatusesUseCase;
         private readonly IInvoicesUseCase _invoicesUseCase;
         private readonly IPayRunUseCase _payRunUseCase;
+        private readonly IGetUserPendingInvoicesUseCase _getUserPendingInvoicesUseCase;
 
-        public InvoicesController(IGetInvoiceItemPaymentStatusesUseCase getInvoiceItemPaymentStatusesUseCase, IInvoicesUseCase invoicesUseCase, IPayRunUseCase payRunUseCase)
+        public InvoicesController(IGetInvoiceItemPaymentStatusesUseCase getInvoiceItemPaymentStatusesUseCase, IInvoicesUseCase invoicesUseCase, IPayRunUseCase payRunUseCase,
+            IGetUserPendingInvoicesUseCase getUserPendingInvoicesUseCase)
         {
             _getInvoiceItemPaymentStatusesUseCase = getInvoiceItemPaymentStatusesUseCase;
             _invoicesUseCase = invoicesUseCase;
             _payRunUseCase = payRunUseCase;
+            _getUserPendingInvoicesUseCase = getUserPendingInvoicesUseCase;
         }
 
         [ProducesResponseType(typeof(IEnumerable<InvoiceItemPaymentStatusResponse>), StatusCodes.Status200OK)]
@@ -79,6 +83,26 @@ namespace LBH.AdultSocialCare.Transactions.Api.V1.Controllers
         {
             var res = await _payRunUseCase.GetHeldInvoicePaymentsUseCase().ConfigureAwait(false);
             return Ok(res);
+        }
+
+        [HttpGet("pending/{serviceUserId}")]
+        [ProducesResponseType(typeof(IEnumerable<PendingInvoicesResponse>), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(string), StatusCodes.Status422UnprocessableEntity)]
+        [ProducesDefaultResponseType]
+        public async Task<ActionResult<IEnumerable<PendingInvoicesResponse>>> GetUserPendingInvoices(Guid serviceUserId)
+        {
+            var result = await _getUserPendingInvoicesUseCase.GetUserPendingInvoices(serviceUserId).ConfigureAwait(false);
+            return Ok(result);
+        }
+
+        [HttpPost]
+        [ProducesResponseType(typeof(IEnumerable<InvoiceResponse>), StatusCodes.Status200OK)]
+        [ProducesDefaultResponseType]
+        public async Task<ActionResult<IEnumerable<InvoiceResponse>>> CreateInvoice([FromBody] InvoiceForCreationRequest invoiceForCreationRequest)
+        {
+            var result = await _invoicesUseCase.CreateInvoiceUseCase(invoiceForCreationRequest.ToDomain()).ConfigureAwait(false);
+            return Ok(result);
         }
     }
 }
